Add docs, cleanup implementations, begin adding animated layer indications
