+++ users/techworker/rgb/config_post.h
@@ -4,22 +4,24 @@
 
 #pragma once
 
+#ifdef RGB_MATRIX_CUSTOM_USER
 // Default LED counts
-#ifndef HOME_LED_COUNT
-#   define HOME_LED_COUNT 0
-#endif
-#ifndef FN_LED_COUNT
-#   define FN_LED_COUNT 0
-#endif
-#ifndef MOD_LED_COUNT
-#   define MOD_LED_COUNT 0
-#endif
-#ifndef ARROW_LED_COUNT
-#   define ARROW_LED_COUNT 0
-#endif
-#ifndef NUMPAD_LED_COUNT
-#   define NUMPAD_LED_COUNT 0
-#endif
+#   ifndef HOME_LED_COUNT
+#      define HOME_LED_COUNT 0
+#   endif
+#   ifndef FN_LED_COUNT
+#      define FN_LED_COUNT 0
+#   endif
+#   ifndef MOD_LED_COUNT
+#      define MOD_LED_COUNT 8
+#   endif
+#   ifndef ARROW_LED_COUNT
+#      define ARROW_LED_COUNT 4
+#   endif
+#   ifndef NUMPAD_LED_COUNT
+#      define NUMPAD_LED_COUNT 10
+#   endif
+#endif // RGB_MATRIX_CUSTOM_USER
 
 #ifndef RGB_MATRIX_REST_MODE
 #    if defined(SPLIT_KEYBOARD) || defined(KEYBOARD_ergodox_ez) || defined(KEYBOARD_zsa)
